requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crispy-forms==1.14.0
django-extensions==3.2.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
